pyproject.toml
[project]
name = "fh-plotly"
version = "0.2.0"
description = "Plotly in FastHTML"
authors = [
    {name = "Carlo Lepelaars", email = "info@carlolepelaars.nl"},
]
license = {text = "MIT License"}
readme = "README.md"
requires-python = ">=3.10,<4"

dependencies = [
    "python-fasthtml>=0.4.0",
    "plotly>=5.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.2",
    "pandas>=2.2.2",
    "numpy>=2.0.1",
    "fastcore>=1.7.1",
]

[tool.ruff]
line-length = 300

[tool.ruff.lint]
ignore = ["I001","F403", "F405", "F811"] 
select = ["E", "F", "I"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
